allseasons/requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django

django-model-utils

django-bootstrap3

django-geoposition==0.3.1  # pip install https://github.com/philippbosch/django-geoposition/archive/django-1.11.zip

django-formtools
